create Toolbar component
