+++ @antv/gatsby-theme-antv/site/components/PlayGround.module.less
@@ -27,57 +27,6 @@
     overflow: hidden;
     height: 100%;
 
-    .toolbar {
-      height: 36px;
-      line-height: 36px;
-      border-bottom: 1px solid @border-color-split;
-      width: 100%;
-      padding: 0 16px;
-      text-align: right;
-      color: #777;
-      background: #f9f9f9;
-
-      > * {
-        display: inline-block;
-        margin-left: 2px;
-        transition: all 0.3s;
-
-        &:hover {
-          color: #222;
-        }
-      }
-
-      .codesandbox {
-        border: 0;
-        background: none;
-        padding: 0;
-        cursor: pointer;
-      }
-
-      .riddle {
-        width: 16px;
-        height: 16px;
-        overflow: hidden;
-        text-indent: -9999px;
-        background: transparent
-          url('https://gw.alipayobjects.com/zos/rmsportal/DlHbxMCyeuyOrqOdbgik.svg')
-          center / 12px no-repeat;
-        border: 0;
-        cursor: pointer;
-        transition: all 0.3s;
-        position: relative;
-        top: -1px;
-      }
-
-      :global(.ant-typography-copy) {
-        color: #777;
-
-        &:hover {
-          color: #222;
-        }
-      }
-    }
-
     .codemirror {
       overflow: hidden;
       height: 100%;
